refactor: 코드스타일 통일
- ApiResponseDto 통일
- 패키지 이름 및 위치 수정
