--- src/main/java/com/example/sixneek/ApiResponseDto.java
@@ -1,15 +1,0 @@
-package com.example.sixneek;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@JsonInclude(value = JsonInclude.Include.NON_NULL)
-public class ApiResponseDto<T> {
-    private Integer status;
-    private String message;
-    private T data;
-}
