+++ src/main/java/com/example/sixneek/global/dto/ApiResponseDto.java
@@ -1,19 +1,20 @@
 package com.example.sixneek.global.dto;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
+import lombok.NoArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonInclude(value = JsonInclude.Include.NON_NULL)
 public class ApiResponseDto<T> {
 
     private HttpStatus status;
     private String message;
     private T data;
-
-    public ApiResponseDto(HttpStatus status, String message, T data) {
-        this.status = status;
-        this.message = message;
-        this.data = data;
-    }
 }
