+++ src/main/java/com/example/sixneek/member/entity/Member.java
@@ -7,8 +7,6 @@ import jakarta.validation.constraints.Email;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-
-
 import java.util.ArrayList;
 import java.util.List;
